form_12/requirements.txt
pandas_gbq==0.15.0
pyarrow==3.0.0
pandas==1.2.4
requests==2.25.1
gspread==4.0.1
gspread_dataframe==3.2.1
protobuf==3.19.0
google-cloud-bigquery
